backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pydantic==2.6.1
pydantic-settings==2.1.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
httpx==0.27.0
apscheduler==3.10.4
alembic==1.13.1
